+++ rtl/core/neorv32_package.vhd
@@ -20,7 +20,7 @@ package neorv32_package is
 
   -- Architecture Constants -----------------------------------------------------------------
   -- -------------------------------------------------------------------------------------------
-  constant hw_version_c  : std_ulogic_vector(31 downto 0) := x"01120800"; -- hardware version
+  constant hw_version_c  : std_ulogic_vector(31 downto 0) := x"01120807"; -- hardware version
   constant int_bus_tmo_c : natural := 16; -- internal bus timeout window; has to be a power of two
   constant alu_cp_tmo_c  : natural := 9;  -- log2 of max ALU co-processor execution cycles
 
@@ -38,11 +38,10 @@ package neorv32_package is
 -- SoC Peripheral/IO Address Space Layout
 -- **********************************************************************************************************
 
-  -- Start of uncached memory access (256MB page / 4 MSBs only) --
-  constant mem_uncached_begin_c : std_ulogic_vector(31 downto 0) := x"f0000000";
+  constant mem_uncached_begin_c : std_ulogic_vector(31 downto 0) := x"f0000000"; -- 256MB page
+  constant mem_io_dev_size_c    : natural := 64*1024; -- size of a single IO device (bytes)
   constant mem_io_base_c        : std_ulogic_vector(31 downto 0) := x"ffe00000";
-  constant mem_io_size_c        : natural := 32*64*1024; -- 32 * iodev_size_c
-  constant iodev_size_c         : natural := 64*1024; -- size of a single IO device (bytes)
+  constant mem_io_size_c        : natural := 32 * mem_io_dev_size_c;
 
   -- IO Address Map (base address must be aligned to the region's size) --
   constant base_io_bootrom_c : std_ulogic_vector(31 downto 0) := x"ffe00000";
@@ -1014,7 +1013,6 @@ package neorv32_package is
       twd_sda_i      : in  std_ulogic := 'H';
       twd_sda_o      : out std_ulogic;
       twd_scl_i      : in  std_ulogic := 'H';
-      twd_scl_o      : out std_ulogic;
       -- 1-Wire Interface (available if IO_ONEWIRE_EN = true) --
       onewire_i      : in  std_ulogic := 'H';
       onewire_o      : out std_ulogic;