+++ rtl/core/neorv32_top.vhd
@@ -223,7 +223,6 @@ entity neorv32_top is
     twd_sda_i      : in  std_ulogic := 'H';                                  -- serial data line sense input
     twd_sda_o      : out std_ulogic;                                         -- serial data line output (pull low only)
     twd_scl_i      : in  std_ulogic := 'H';                                  -- serial clock line sense input
-    twd_scl_o      : out std_ulogic;                                         -- serial clock line output (pull low only)
 
     -- 1-Wire Interface (available if IO_ONEWIRE_EN = true) --
     onewire_i      : in  std_ulogic := 'H';                                  -- 1-wire bus sense input
@@ -458,11 +457,11 @@ begin
   -- **************************************************************************************************************************
 
   -- fast interrupt requests (FIRQs) --
-  cpu_firq(0)  <= firq(FIRQ_TWD); -- highest priority
+  cpu_firq(0)  <= '0'; -- reserved
   cpu_firq(1)  <= firq(FIRQ_CFS);
   cpu_firq(2)  <= firq(FIRQ_UART0);
   cpu_firq(3)  <= firq(FIRQ_UART1);
-  cpu_firq(4)  <= '0'; -- reserved
+  cpu_firq(4)  <= firq(FIRQ_TWD);
   cpu_firq(5)  <= firq(FIRQ_TRACER);
   cpu_firq(6)  <= firq(FIRQ_SPI);
   cpu_firq(7)  <= firq(FIRQ_TWI);
@@ -473,7 +472,7 @@ begin
   cpu_firq(12) <= firq(FIRQ_GPTMR);
   cpu_firq(13) <= firq(FIRQ_ONEWIRE);
   cpu_firq(14) <= firq(FIRQ_SLINK);
-  cpu_firq(15) <= firq(FIRQ_TRNG); -- lowest priority
+  cpu_firq(15) <= firq(FIRQ_TRNG);
 
   -- CPU core(s) + optional caches + bus switch --
   core_complex_gen:
@@ -921,39 +920,39 @@ begin
     -- -------------------------------------------------------------------------------------------
     neorv32_bus_io_switch_inst: entity neorv32.neorv32_bus_io_switch
     generic map (
-      DEV_SIZE  => iodev_size_c,
-      DEV_00_EN => bootrom_en_c,  DEV_00_BASE => base_io_bootrom_c,
-      DEV_01_EN => false,         DEV_01_BASE => (others => '0'), -- reserved
-      DEV_02_EN => false,         DEV_02_BASE => (others => '0'), -- reserved
-      DEV_03_EN => false,         DEV_03_BASE => (others => '0'), -- reserved
-      DEV_04_EN => false,         DEV_04_BASE => (others => '0'), -- reserved
-      DEV_05_EN => false,         DEV_05_BASE => (others => '0'), -- reserved
-      DEV_06_EN => false,         DEV_06_BASE => (others => '0'), -- reserved
-      DEV_07_EN => false,         DEV_07_BASE => (others => '0'), -- reserved
-      DEV_08_EN => false,         DEV_08_BASE => (others => '0'), -- reserved
-      DEV_09_EN => false,         DEV_09_BASE => (others => '0'), -- reserved
-      DEV_10_EN => IO_TWD_EN,     DEV_10_BASE => base_io_twd_c,
-      DEV_11_EN => IO_CFS_EN,     DEV_11_BASE => base_io_cfs_c,
-      DEV_12_EN => IO_SLINK_EN,   DEV_12_BASE => base_io_slink_c,
-      DEV_13_EN => IO_DMA_EN,     DEV_13_BASE => base_io_dma_c,
-      DEV_14_EN => false,         DEV_14_BASE => (others => '0'), -- reserved
-      DEV_15_EN => false,         DEV_15_BASE => (others => '0'), -- reserved
-      DEV_16_EN => io_pwm_en_c,   DEV_16_BASE => base_io_pwm_c,
-      DEV_17_EN => io_gptmr_en_c, DEV_17_BASE => base_io_gptmr_c,
-      DEV_18_EN => IO_ONEWIRE_EN, DEV_18_BASE => base_io_onewire_c,
-      DEV_19_EN => IO_TRACER_EN,  DEV_19_BASE => base_io_tracer_c,
-      DEV_20_EN => IO_CLINT_EN,   DEV_20_BASE => base_io_clint_c,
-      DEV_21_EN => IO_UART0_EN,   DEV_21_BASE => base_io_uart0_c,
-      DEV_22_EN => IO_UART1_EN,   DEV_22_BASE => base_io_uart1_c,
-      DEV_23_EN => IO_SDI_EN,     DEV_23_BASE => base_io_sdi_c,
-      DEV_24_EN => IO_SPI_EN,     DEV_24_BASE => base_io_spi_c,
-      DEV_25_EN => IO_TWI_EN,     DEV_25_BASE => base_io_twi_c,
-      DEV_26_EN => IO_TRNG_EN,    DEV_26_BASE => base_io_trng_c,
-      DEV_27_EN => IO_WDT_EN,     DEV_27_BASE => base_io_wdt_c,
-      DEV_28_EN => io_gpio_en_c,  DEV_28_BASE => base_io_gpio_c,
-      DEV_29_EN => IO_NEOLED_EN,  DEV_29_BASE => base_io_neoled_c,
-      DEV_30_EN => true,          DEV_30_BASE => base_io_sysinfo_c, -- allways enabled
-      DEV_31_EN => OCD_EN,        DEV_31_BASE => base_io_ocd_c
+      DEV_SIZE  => mem_io_dev_size_c,
+      DEV_00_EN => bootrom_en_c,      DEV_00_BASE => base_io_bootrom_c,
+      DEV_01_EN => false,             DEV_01_BASE => (others => '0'), -- reserved
+      DEV_02_EN => false,             DEV_02_BASE => (others => '0'), -- reserved
+      DEV_03_EN => false,             DEV_03_BASE => (others => '0'), -- reserved
+      DEV_04_EN => false,             DEV_04_BASE => (others => '0'), -- reserved
+      DEV_05_EN => false,             DEV_05_BASE => (others => '0'), -- reserved
+      DEV_06_EN => false,             DEV_06_BASE => (others => '0'), -- reserved
+      DEV_07_EN => false,             DEV_07_BASE => (others => '0'), -- reserved
+      DEV_08_EN => false,             DEV_08_BASE => (others => '0'), -- reserved
+      DEV_09_EN => false,             DEV_09_BASE => (others => '0'), -- reserved
+      DEV_10_EN => IO_TWD_EN,         DEV_10_BASE => base_io_twd_c,
+      DEV_11_EN => IO_CFS_EN,         DEV_11_BASE => base_io_cfs_c,
+      DEV_12_EN => IO_SLINK_EN,       DEV_12_BASE => base_io_slink_c,
+      DEV_13_EN => IO_DMA_EN,         DEV_13_BASE => base_io_dma_c,
+      DEV_14_EN => false,             DEV_14_BASE => (others => '0'), -- reserved
+      DEV_15_EN => false,             DEV_15_BASE => (others => '0'), -- reserved
+      DEV_16_EN => io_pwm_en_c,       DEV_16_BASE => base_io_pwm_c,
+      DEV_17_EN => io_gptmr_en_c,     DEV_17_BASE => base_io_gptmr_c,
+      DEV_18_EN => IO_ONEWIRE_EN,     DEV_18_BASE => base_io_onewire_c,
+      DEV_19_EN => IO_TRACER_EN,      DEV_19_BASE => base_io_tracer_c,
+      DEV_20_EN => IO_CLINT_EN,       DEV_20_BASE => base_io_clint_c,
+      DEV_21_EN => IO_UART0_EN,       DEV_21_BASE => base_io_uart0_c,
+      DEV_22_EN => IO_UART1_EN,       DEV_22_BASE => base_io_uart1_c,
+      DEV_23_EN => IO_SDI_EN,         DEV_23_BASE => base_io_sdi_c,
+      DEV_24_EN => IO_SPI_EN,         DEV_24_BASE => base_io_spi_c,
+      DEV_25_EN => IO_TWI_EN,         DEV_25_BASE => base_io_twi_c,
+      DEV_26_EN => IO_TRNG_EN,        DEV_26_BASE => base_io_trng_c,
+      DEV_27_EN => IO_WDT_EN,         DEV_27_BASE => base_io_wdt_c,
+      DEV_28_EN => io_gpio_en_c,      DEV_28_BASE => base_io_gpio_c,
+      DEV_29_EN => IO_NEOLED_EN,      DEV_29_BASE => base_io_neoled_c,
+      DEV_30_EN => true,              DEV_30_BASE => base_io_sysinfo_c, -- allways enabled
+      DEV_31_EN => OCD_EN,            DEV_31_BASE => base_io_ocd_c
     )
     port map (
       clk_i        => clk_i,
@@ -1280,7 +1279,6 @@ begin
         twd_sda_i => twd_sda_i,
         twd_sda_o => twd_sda_o,
         twd_scl_i => twd_scl_i,
-        twd_scl_o => twd_scl_o,
         irq_o     => firq(FIRQ_TWD)
       );
     end generate;
@@ -1289,7 +1287,6 @@ begin
     if not IO_TWD_EN generate
       iodev_rsp(IODEV_TWD) <= rsp_terminate_c;
       twd_sda_o            <= '1';
-      twd_scl_o            <= '1';
       firq(FIRQ_TWD)       <= '0';
     end generate;