+++ sim/neorv32_tb.vhd
@@ -93,7 +93,7 @@ architecture neorv32_tb_rtl of neorv32_tb is
   signal gpio : std_ulogic_vector(31 downto 0);
   signal i2c_scl, i2c_sda : std_logic;
   signal twi_scl_i, twi_scl_o, twi_sda_i, twi_sda_o : std_ulogic;
-  signal twd_scl_i, twd_scl_o, twd_sda_i, twd_sda_o : std_ulogic;
+  signal twd_scl_i, twd_sda_i, twd_sda_o : std_ulogic;
   signal onewire : std_logic;
   signal onewire_i, onewire_o : std_ulogic;
   signal spi_csn : std_ulogic_vector(7 downto 0);
@@ -306,10 +306,9 @@ begin
     twi_scl_i => twi_scl_i,
     twi_scl_o => twi_scl_o,
     -- TWD --
-    twd_sda_i => twd_sda_i,
-    twd_sda_o => twd_sda_o,
-    twd_scl_i => twd_scl_i,
-    twd_scl_o => twd_scl_o,
+    twd_sda_i      => twd_sda_i,
+    twd_sda_o      => twd_sda_o,
+    twd_scl_i      => twd_scl_i,
     -- 1-Wire Interface --
     onewire_i => onewire_i,
     onewire_o => onewire_o,
@@ -328,7 +327,8 @@ begin
     irq_mei_i      => mei
   );
 
-  -- Two-Wire Bus - Tri-State Drivers (modules can only actively pull the signals low) ------
+
+  -- Two-Wire Bus - Tri-State Drivers (modules can only actively pull low) ------------------
   -- -------------------------------------------------------------------------------------------
   i2c_sda <= '0' when (twi_sda_o = '0') else
              'Z';
@@ -337,10 +337,7 @@ begin
   twi_sda_i <= std_ulogic(i2c_sda); -- sense input
   twi_scl_i <= std_ulogic(i2c_scl); -- sense input
 
-  i2c_sda <= '0' when (twd_sda_o = '0') else
-             'Z';
-  i2c_scl <= '0' when (twd_scl_o = '0') else
-             'Z';
+  i2c_sda   <= '0' when (twd_sda_o = '0') else 'Z';
   twd_sda_i <= std_ulogic(i2c_sda); -- sense input
   twd_scl_i <= std_ulogic(i2c_scl); -- sense input
 
@@ -348,7 +345,8 @@ begin
   i2c_scl <= 'H';
   i2c_sda <= 'H';
 
-  -- One-Wire Bus - Tri-State Driver (module can only actively pull the signals low) --------
+
+  -- One-Wire Bus - Tri-State Driver (module can only actively pull low) --------------------
   -- -------------------------------------------------------------------------------------------
   onewire <= '0' when (onewire_o = '0') else
              'Z';