+++ sw/image_gen/image_gen.c
@@ -12,53 +12,95 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#ifdef __APPLE__
+#include <libelf.h>
+#else
+#include <elf.h>
+#endif
 
-// executable signature ("magic word")
-const uint32_t signature = 0xB007C0DE;
+// executable signature identifier ("magic word", for bootloader only)
+const uint32_t signature_c = 0x214F454E;
 
-// output file types (operation select)
+// output image types (operation select)
 enum operation_enum {
   OP_EXE,
   OP_VHD,
   OP_BIN,
-  OP_HEX,
   OP_COE,
   OP_MEM,
   OP_MIF
 };
 
+// bootloader executable header
+typedef struct __attribute__((packed,aligned(4))) {
+  uint32_t signature;
+  uint32_t base_addr;
+  uint32_t size;
+  uint32_t checksum;
+} exe_header_t;
+
+// ************************************************************
+// Write 32-bit data to file (little-Endian).
+// ************************************************************
+void write32(uint32_t d, FILE *f) {
+
+  fputc((unsigned char)((d >>  0) & 0xFF), f);
+  fputc((unsigned char)((d >>  8) & 0xFF), f);
+  fputc((unsigned char)((d >> 16) & 0xFF), f);
+  fputc((unsigned char)((d >> 24) & 0xFF), f);
+}
+
+// ************************************************************
+// Read ELF section.
+// ************************************************************
+void *read_section(FILE *f, Elf32_Shdr *sh) {
+
+  void *data = malloc(sh->sh_size);
+  fseek(f, sh->sh_offset, SEEK_SET);
+  if (fread(data, 1, sh->sh_size, f) <= 0) {
+    return NULL;
+  }
+  else {
+    return data;
+  }
+}
+
+// ************************************************************
+// Show help menu.
+// ************************************************************
 void print_help(void){
+
   printf(
     "NEORV32 executable image generator\n"
     "\n"
     "Usage:    image_gen [options]\n"
-    "Example:  image_gen -i main.bin -o main_exe.bin -t exe\n"
+    "Example:  image_gen -i main.elf -o main_exe.bin -t exe\n"
     "\n"
     "Options:\n"
     "  -h            Show this help text and exit\n"
-    "  -i file_name  Input binary file name; mandatory\n"
-    "  -o file_name  Output file name; mandatory\n"
-    "  -t format     Output image format; default is 'exe'\n"
-    "  -n number     Number of bytes per line (hex, coe, mem, mif); default is 4\n"
+    "  -i file_name  ELF input file\n"
+    "  -o file_name  Image output file\n"
+    "  -t format     Image output format\n"
     "\n"
     "Image formats (using little-Endian byte ordering):\n"
     "  exe  Executable for bootloader upload (binary file with header) \n"
     "  vhd  VHDL memory image (raw executable)\n"
     "  bin  Binary file (raw executable)\n"
-    "  hex  HEX file ('number' bytes per line, ASCII, raw executable)\n"
-    "  coe  COE file ('number' bytes per line, ASCII, raw executable)\n"
-    "  mem  MEM file ('number' bytes per line, ASCII, raw executable)\n"
-    "  mif  MIF file ('number' bytes per line, ASCII, raw executable)\n"
+    "  coe  COE file (8x hex per line, ASCII, raw executable)\n"
+    "  mem  MEM file (8x hex per line, ASCII, raw executable)\n"
+    "  mif  MIF file (8x hex per line, ASCII, raw executable)\n"
   );
 }
 
+// ************************************************************
+// Main.
+// ************************************************************
 int main(int argc, char *argv[]) {
 
   FILE *input = NULL, *output = NULL;
   char *input_file = NULL, *output_file = NULL, tmp_string[1024];
-  uint32_t u32 = 0, checksum = 0;
+  uint32_t checksum = 0;
   unsigned int i = 0, operation = OP_EXE, raw_exe_size = 0, ext_exe_size = 0;
-  unsigned char byte = 0;
 
   // show help menu if there are no arguments
   if (argc <= 1) {
@@ -66,7 +108,9 @@ int main(int argc, char *argv[]) {
     return 0;
   }
 
+  // ****************************************
   // parse arguments
+  // ****************************************
   for (i = 1; i < argc; i++) {
     // show help
     if (strcmp(argv[i], "-h") == 0) {
@@ -87,7 +131,6 @@ int main(int argc, char *argv[]) {
       if      (strcmp(argv[i], "exe") == 0) { operation = OP_EXE; }
       else if (strcmp(argv[i], "vhd") == 0) { operation = OP_VHD; }
       else if (strcmp(argv[i], "bin") == 0) { operation = OP_BIN; }
-      else if (strcmp(argv[i], "hex") == 0) { operation = OP_HEX; }
       else if (strcmp(argv[i], "coe") == 0) { operation = OP_COE; }
       else if (strcmp(argv[i], "mem") == 0) { operation = OP_MEM; }
       else if (strcmp(argv[i], "mif") == 0) { operation = OP_MIF; }
@@ -96,31 +139,105 @@ int main(int argc, char *argv[]) {
         return -1;
       }
     }
-    // unknown
+    // invalid
     else {
-      printf("[ERROR] Unknown flag '%s'!\n", argv[i]);
+      printf("[ERROR] Invalid flag '%s'!\n", argv[i]);
       return -1;
     }
   }
 
-  // open input file
+  // ****************************************
+  // open input/output files
+  // ****************************************
   input = fopen(input_file, "rb");
   if (input == NULL) {
     printf("[ERROR] Input file error (%s)!\n", input_file);
     return -2;
   }
 
-  // get input file size
-  fseek(input, 0L, SEEK_END);
-  raw_exe_size = (unsigned int)ftell(input);
-  rewind(input);
-  if (raw_exe_size == 0) {// input file empty?
+  output = fopen(output_file, "wb");
+  if (output == NULL) {
+    printf("[ERROR] Output file error (%s)!\n", output_file);
+    fclose(output);
+    return -2;
+  }
+
+  // ****************************************
+  // parse ELF
+  // ****************************************
+
+  Elf32_Ehdr elf;
+  if (fread(&elf, 1, sizeof(elf), input) <= 0) {
     printf("[ERROR] Input file is empty (%s)!\n", input_file);
-    fclose(input);
+    return -2;
+  }
+
+  if (memcmp(elf.e_ident, ELFMAG, SELFMAG) != 0) {
+    printf("[ERROR] Input file is not an ELF (%s)!\n", input_file);
+    return -2;
+  }
+
+  // base address (= entry point)
+  uint32_t base_addr = (uint32_t)elf.e_entry;
+
+  // section header
+  Elf32_Shdr *shdrs = malloc(elf.e_shentsize * elf.e_shnum);
+  if (!shdrs) {
+    printf("[ERROR] malloc failed!\n");
+    return -1;
+  }
+  fseek(input, elf.e_shoff, SEEK_SET);
+  if (fread(shdrs, elf.e_shentsize, elf.e_shnum, input) <= 0) {
+    printf("[ERROR] Input file read error (%s)!\n", input_file);
+    return -2;
+  }
+
+  // section string table
+  Elf32_Shdr shstr = shdrs[elf.e_shstrndx];
+  char *shstrtab = read_section(input, &shstr);
+  void *text = NULL;
+  void *rodata = NULL;
+  void *data = NULL;
+  unsigned int text_size = 0;
+  unsigned int rodata_size = 0;
+  unsigned int data_size = 0;
+
+  // scan section headers
+  for (i = 0; i < elf.e_shnum; i++) {
+    const char *section_name = shstrtab + shdrs[i].sh_name;
+    if (strcmp(section_name, ".text") == 0) {
+      text = read_section(input, &shdrs[i]);
+      text_size = (unsigned int)shdrs[i].sh_size;
+    }
+    if (strcmp(section_name, ".rodata") == 0) {
+      rodata = read_section(input, &shdrs[i]);
+      rodata_size = (unsigned int)shdrs[i].sh_size;
+    }
+    if (strcmp(section_name, ".data") == 0) {
+      data = read_section(input, &shdrs[i]);
+      data_size = (unsigned int)shdrs[i].sh_size;
+    }
+  }
+
+  fclose(input);
+
+  // ****************************************
+  // generate raw image
+  // ****************************************
+
+  // debug
+//printf(".text:   %d bytes\n", text_size);
+//printf(".rodata: %d bytes\n", rodata_size);
+//printf(".data:   %d bytes\n", data_size);
+
+  // final image size
+  raw_exe_size = text_size + rodata_size + data_size;
+  if (raw_exe_size == 0) {// input file empty?
+    printf("[ERROR] Image is empty!\n");
     return -2;
   }
   if ((raw_exe_size % 4) != 0) {
-    printf("[WARNING] Input image size is not a multiple of 4 bytes!\n");
+    printf("[WARNING] Image size is not a multiple of 4 bytes!\n");
   }
 
   // make sure memory array is a power of two
@@ -129,52 +246,50 @@ int main(int argc, char *argv[]) {
     ext_exe_size *= 2;
   }
 
-  // open output file
-  output = fopen(output_file, "wb");
-  if (output == NULL) {
-    printf("[ERROR] Output file error (%s)!\n", output_file);
-    fclose(output);
-    return -2;
+  // construct raw image
+  uint8_t *raw_image = malloc(raw_exe_size);
+  uint32_t *raw_image32 = (uint32_t *)raw_image;
+  if (!raw_image) {
+    printf("[ERROR] malloc failed!\n");
+    return -1;
   }
+  memcpy(raw_image,                           text,   text_size);   // start with .text
+  memcpy(raw_image + text_size,               rodata, rodata_size); // append .rodata
+  memcpy(raw_image + text_size + rodata_size, data,   data_size);   // append .data
 
   // --------------------------------------------------------------------------
   // executable for bootloader upload (including header)
   // --------------------------------------------------------------------------
   if (operation == OP_EXE) {
 
+    exe_header_t header;
+
     // reserve header space
-    for (i=0; i<12; i++) {
+    for (i = 0; i < sizeof(header); i++) {
       fputc(0, output);
     }
 
-    // actual data
+    // actual data and checksum
     checksum = 0;
-    rewind(input);
-    while(fread(&u32, sizeof(uint32_t), 1, input) != 0) {
-      checksum += u32; // checksum: sum complement
-      fputc((unsigned char)((u32 >>  0) & 0xFF), output);
-      fputc((unsigned char)((u32 >>  8) & 0xFF), output);
-      fputc((unsigned char)((u32 >> 16) & 0xFF), output);
-      fputc((unsigned char)((u32 >> 24) & 0xFF), output);
+    for (i = 0; i < raw_exe_size/4; i++) {
+      checksum += raw_image32[i];
+      write32(raw_image32[i], output);
     }
 
+    // setup header
     rewind(output);
-    // header: signature
-    fputc((unsigned char)((signature >>  0) & 0xFF), output);
-    fputc((unsigned char)((signature >>  8) & 0xFF), output);
-    fputc((unsigned char)((signature >> 16) & 0xFF), output);
-    fputc((unsigned char)((signature >> 24) & 0xFF), output);
-    // header: size
-    fputc((unsigned char)((raw_exe_size >>  0) & 0xFF), output);
-    fputc((unsigned char)((raw_exe_size >>  8) & 0xFF), output);
-    fputc((unsigned char)((raw_exe_size >> 16) & 0xFF), output);
-    fputc((unsigned char)((raw_exe_size >> 24) & 0xFF), output);
-    // header: checksum (sum complement)
-    checksum = ~checksum;
-    fputc((unsigned char)((checksum >>  0) & 0xFF), output);
-    fputc((unsigned char)((checksum >>  8) & 0xFF), output);
-    fputc((unsigned char)((checksum >> 16) & 0xFF), output);
-    fputc((unsigned char)((checksum >> 24) & 0xFF), output);
+    header.signature = signature_c;
+    header.base_addr = base_addr;
+    header.size      = raw_exe_size;
+    header.checksum  = ~checksum;
+    write32(header.signature, output);
+    write32(header.base_addr, output);
+    write32(header.size, output);
+    write32(header.checksum, output);
+
+    // report
+    printf("Executable (EXE): %d bytes @ 0x%08X, checksum = 0x%08X\n",
+           (unsigned int)header.size, (unsigned int)header.base_addr, (unsigned int)header.checksum);
   }
 
   // --------------------------------------------------------------------------
@@ -217,15 +332,10 @@ int main(int argc, char *argv[]) {
       pkg_name, (ext_exe_size/4)-1, raw_exe_size);
     fputs(tmp_string, output);
 
-    for (i=0; i<(raw_exe_size/4); i++) {
-      if (fread(&u32, sizeof(uint32_t), 1, input) != 0) {
-        snprintf(tmp_string, sizeof(tmp_string), "x\"%08x\",\n", (unsigned int)u32);
-        fputs(tmp_string, output);
-      }
-      else {
-        printf("[WARNING] Unexpected input file end!\n");
-        break;
-      }
+    // data
+    for (i = 0; i < raw_exe_size/4; i++) {
+      snprintf(tmp_string, sizeof(tmp_string), "x\"%08x\",\n", (unsigned int)raw_image32[i]);
+      fputs(tmp_string, output);
     }
 
     // end
@@ -235,27 +345,22 @@ int main(int argc, char *argv[]) {
       "\n"
       "end %s;\n", pkg_name);
     fputs(tmp_string, output);
+
+    // report
+    printf("Executable (VHD): %d bytes\n", raw_exe_size);
   }
 
   // --------------------------------------------------------------------------
-  // executable binary file
+  // executable plain-binary file
   // --------------------------------------------------------------------------
   else if (operation == OP_BIN) {
 
-    while(fread(&byte, sizeof(unsigned char), 1, input) != 0) {
-      fputc(byte, output);
+    for (i = 0; i < raw_exe_size; i++) {
+      fputc((unsigned char)(raw_image[i]), output);
     }
-  }
 
-  // --------------------------------------------------------------------------
-  // executable ASCII hex file
-  // --------------------------------------------------------------------------
-  else if (operation == OP_HEX) {
-
-    while(fread(&u32, sizeof(uint32_t), 1, input) != 0) {
-      snprintf(tmp_string, sizeof(tmp_string), "%08x\n", (unsigned int)u32);
-      fputs(tmp_string, output);
-    }
+    // report
+    printf("Executable (BIN): %d bytes\n", raw_exe_size);
   }
 
   // --------------------------------------------------------------------------
@@ -269,17 +374,18 @@ int main(int argc, char *argv[]) {
     snprintf(tmp_string, sizeof(tmp_string), "memory_initialization_vector=\n");
     fputs(tmp_string, output);
 
-    i = 0;
-    while(fread(&u32, sizeof(uint32_t), 1, input) != 0) {
+    for (i = 0; i < raw_exe_size/4; i++) {
       if (i == ((raw_exe_size/4)-1)) {
-        snprintf(tmp_string, sizeof(tmp_string), "%08x;\n", (unsigned int)u32);
+        snprintf(tmp_string, sizeof(tmp_string), "%08x;\n", (unsigned int)raw_image32[i]);
       }
       else {
-        snprintf(tmp_string, sizeof(tmp_string), "%08x,\n", (unsigned int)u32);
+        snprintf(tmp_string, sizeof(tmp_string), "%08x,\n", (unsigned int)raw_image32[i]);
       }
       fputs(tmp_string, output);
-      i++;
     }
+
+    // report
+    printf("Executable (COE): %d bytes\n", raw_exe_size);
   }
 
   // --------------------------------------------------------------------------
@@ -287,12 +393,13 @@ int main(int argc, char *argv[]) {
   // --------------------------------------------------------------------------
   else if (operation == OP_MEM) {
 
-    i = 0;
-    while(fread(&u32, sizeof(uint32_t), 1, input) != 0) {
-      snprintf(tmp_string, sizeof(tmp_string), "@%08x %08x\n", (unsigned int)i, (unsigned int)u32);
+    for (i = 0; i < raw_exe_size/4; i++) {
+      snprintf(tmp_string, sizeof(tmp_string), "@%08x %08x\n", (unsigned int)i, (unsigned int)raw_image32[i]);
       fputs(tmp_string, output);
-      i++;
     }
+
+    // report
+    printf("Executable (MEM): %d bytes\n", raw_exe_size);
   }
 
   // --------------------------------------------------------------------------
@@ -314,28 +421,32 @@ int main(int argc, char *argv[]) {
     fputs(tmp_string, output);
     snprintf(tmp_string, sizeof(tmp_string), "BEGIN\n");
     fputs(tmp_string, output);
-    i = 0;
-    while(fread(&u32, sizeof(uint32_t), 1, input) != 0) {
-      snprintf(tmp_string, sizeof(tmp_string), "%08x : %08x;\n", (unsigned int)i, (unsigned int)u32);
+
+    for (i = 0; i < raw_exe_size/4; i++) {
+      snprintf(tmp_string, sizeof(tmp_string), "%08x : %08x;\n", (unsigned int)i, (unsigned int)raw_image32[i]);
       fputs(tmp_string, output);
-      i++;
     }
 
     // footer
     snprintf(tmp_string, sizeof(tmp_string), "END;\n");
     fputs(tmp_string, output);
+
+    // report
+    printf("Executable (MIF): %d bytes\n", raw_exe_size);
   }
 
   // invalid operation
   else {
     printf("[ERROR] Invalid operation!\n");
-    fclose(input);
+    free(raw_image);
+    free(shdrs);
     fclose(output);
     return -1;
   }
 
   // clean up
-  fclose(input);
+  free(raw_image);
+  free(shdrs);
   fclose(output);
 
   return 0;